Reorganize RequestLoanForm test files
